==> Scripts/03 Get Leads Addresses.ipynb <==
import keys
import numpy as np
import pandas as pd
import json
import urllib
import requests
import re
import matplotlib.pyplot as plt
import requests
from pandas.io.json import json_normalize
from bs4 import BeautifulSoup
import requests
from pprint import pprint
import json
%matplotlib inline

#df2 = pd.read_csv('leads.csv')

lead['lat'] = np.nan
lead['lng'] = np.nan

df2['address'] = df2['Street'] + ' ' + df2['City'] + ' ' + \
    df2['State/Province'] + ' ' + df2['Zip/Postal Code']  # +' '+ df2['Country']

df2[(df2['lat'].isnull() == True) &
    (df2['address'].isnull() == False) &
    (df2['Zip/Postal Code'].isnull() == False)
    ].head()

#key = 'xxx'
key = keys.google
counter = 0
for i, row in df2[(df2['lat'].isnull() == True) &
                  (df2['address'].isnull() == False) &
                  (df2['Zip/Postal Code'].isnull() == False)
                  ].iterrows():
    address = row.address.replace(' ', '+')
    url = 'https://maps.googleapis.com/maps/api/geocode/json?address='+address+'=&key='+key
    # print(url)
    # print(row.address)
    response = requests.get(url)
    try:
        data = json.loads(response.text)
        lat = data['results'][0]['geometry']['location']['lat']
        lng = data['results'][0]['geometry']['location']['lng']
        df2.loc[df2['Lead ID'] == row['Lead ID'], 'lat'] = lat
        df2.loc[df2['Lead ID'] == row['Lead ID'], 'lng'] = lng
    except:
        #print ('request failure for row', i, row['Unnamed: 0'], data, url)
        pass
    counter += 1
    if counter % 500 == 0:
        print(counter, address, lat, lng)
    if counter % 500 == 0:
        df2.to_csv('lead geocoded.csv', index=False)
        # break

df2.to_csv('lead geocoded.csv', index=False)

print(df2[df2.lat.isnull() == True].shape)
print(df2[df2.lat.isnull() == False].shape)

lead.to_csv('lead geocoded.csv')

lead[(lead['lat'].isnull()) & (lead['lng'].isnull())].shape

lead[lead['lat'].isnull() == False].shape

df2 = pd.read_csv('lead geocoded.csv')

df2.head(100)

print(df2[df2.lat.isnull() == True].shape)
print(df2[df2.lat.isnull() == False].shape)

Many of the address without lats and lngs had faulty coiuntry information; these were corrected manually

df2 = pd.read_csv('./sianedit_leads_with_distances.csv', encoding='ISO-8859-1')

df2.head()

df2.drop(['Unnamed: 0', 'City_1', 'lat_lng', 'closest_ts', 'within_one_tenth',
          'within_one_half', 'within_one', 'within_five'], axis=1, inplace=True)

print(df2[df2.lat.isnull() == True].shape)
print(df2[df2.lat.isnull() == False].shape)

print(df2[df2.address.isnull() == True].shape)
print(df2[df2.address.isnull() == False].shape)

# rebuild address column
df2['address'] = df2['Street'] + ' ' + df2['City'] + ' ' + \
    df2['State/Province'] + ' ' + df2['Zip/Postal Code']+' ' + df2['Country']

print(df2[df2.address.isnull() == True].shape)
print(df2[df2.address.isnull() == False].shape)

# df2[df2.address.isnull()==True].head(100)

print(df2[df2.address.isnull() == True].shape)
print(df2[df2.address.isnull() == False].shape)

print(df2[df2.lat.isnull() == True].shape)
print(df2[df2.lat.isnull() == False].shape)

final_leads = df2[df2.lat.isnull() == False]

final_leads.to_csv('final_leads.csv', index_label=False)

# get transitscreen info

df = pd.read_csv('final_leads.csv')

df['ts_response'] = np.nan
df['mobility_score'] = np.nan
df['carshare'] = np.nan
df['bikeshare'] = np.nan
df['ridehailing'] = np.nan
df['masstransit'] = np.nan

df.head()

base_url = 'http://mobilityscore-dev.transitscreen.io/api/v1/locations.json?&debug=true'
key = keys.ts

for i, row in df[(df['ts_response'].isnull() == True)
                 #& (df['id']=='0061a00000GcQdL') #for test
                 ].iterrows():
    #print(row.id, row.name, row.lat, row.lng)
    try:
        url = base_url + '&key='+key+'&geojson=true&coordinates=' + \
            str(row.lat)+','+str(row.lng)
        #(print(url))
        response = requests.get(url)  # wait up to 10 seconds
        data = json.loads(response.text)
        full_response = data['data']
        mobility_score = full_response['mobilityScore']['score']
        #print (mobility_score)
        raw_scores = full_response['mobilityScore']['rawScoreBreakdown']
        #print (raw_scores)
        carshare = raw_scores['carshare']
        bikeshare = raw_scores['bikeshare']
        ridehailing = raw_scores['ridehailing']
        masstransit = raw_scores['masstransit']
        #print(carshare, bikeshare, ridehailing, masstransit)
        df.loc[df['id'] == row.id, 'ts_response'] = data
        df.loc[df['id'] == row.id, 'mobility_score'] = mobility_score
        df.loc[df['id'] == row.id, 'carshare'] = carshare
        df.loc[df['id'] == row.id, 'bikeshare'] = bikeshare
        df.loc[df['id'] == row.id, 'ridehailing'] = ridehailing
        df.loc[df['id'] == row.id, 'masstransit'] = masstransit
    except:
        print('error on', i, raw_scores)
    if i % 100 == 0:
        df.to_csv('model_data.csv', index=False)
        print(i, raw_scores)
