==> Scripts/04 scrape for missing addresses.ipynb <==
import pandas as pd
import numpy as np
import requests
from bs4 import BeautifulSoup
from selenium import webdriver
import sys
import re
import webbrowser
from time import sleep

# https://stackoverflow.com/questions/4028904/how-to-get-the-home-directory-in-python
from os.path import expanduser
from os import listdir
executable_path = expanduser("~") + "/chromedriver"

if 'chromedriver.exe' not in listdir(expanduser("~")):
    print('chomedriver.exe not found in the home directory! Refer to Selenium docs.')
    sys.exit()

driver = webdriver.Chrome(executable_path=executable_path)

from selenium import webdriver
from bs4 import BeautifulSoup

import pandas as pd


# https://stackoverflow.com/questions/4028904/how-to-get-the-home-directory-in-python
from os.path import expanduser
from os import listdir
executable_path = expanduser("~") + "/chromedriver"

if 'chromedriver.exe' not in listdir(expanduser("~")):
    print('chomedriver.exe not found in the home directory! Refer to Selenium docs.')
    sys.exit()

driver = webdriver.Chrome(executable_path=executable_path)

#opportunities = pd.read_csv('opportunities.csv')
# use this if restarting scraper
opportunities = pd.read_csv('opportunities_scraped.csv')

opportunities.head()

# opportunities['scraped_address']=np.nan

opportunities[opportunities.scraped_address != 'not found'].head()

opportunities.head(33)

driver = webdriver.Chrome(
    executable_path="/Users/rajchakrabarty/Downloads/chromedriver")

# export to csv
opportunities.to_csv('opportunities_scraped.csv', index_label=False)

def scrape():
    driver = webdriver.Chrome(
        executable_path="/Users/rajchakrabarty/Downloads/chromedriver")
    counter = 0
    for i, row in opportunities[(opportunities['scraped_address'].isnull() == True)].iterrows():
        search_term = ''
        url = ''
        address = ''
        search_term = row['Opportunity Name'].replace(' ', '+')+'+address'
        url = 'https://www.google.com/search?\&q='+search_term
        driver.get(url)
        search_bar = driver.find_element_by_id('lst-ib')
        search_bar.clear()
        search_bar.send_keys(search_term)
        response = driver.page_source
        html = BeautifulSoup(response, 'lxml')
        span = html.find('span', {'class': 'LrzXr'})
        #print (i, search_term)
        try:
            address = span.text
        except:
            address = 'not found'
        # print(address)
        opportunities.loc[opportunities['Opportunity ID'] ==
                          row['Opportunity ID'], 'scraped_address'] = address
        counter += 1
        if counter % 20 == 0:
            print(i, search_term)
            print(address)
            # sleep(10)


print('start')
keep_going = True
while keep_going:
    try:
        scrape()
        keep_going = False
    except:
        opportunities.to_csv('opportunities_scraped.csv', index_label=False)

opportunities.to_csv('opportunities_scraped.csv', index_label=False)
